batch: Move presence placement choices
Presence constraints now share missing-claim detection through a helper
module, and merge candidate enumeration still builds reviewed options for
unresolved claimed-line placement.

The placement-choice value object and ambiguity helpers still live in
presence realization. That makes reviewed candidate discovery depend on the
constraint application module for data that only describes where a missing
run can fit.

This commit continues the presence placement split by moving choice
enumeration, ambiguity keys, and target-span calculation into a presence
placement choices module.

Subsequent commits will record that placement-choice ownership in
architecture tests.
